+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes for Mollie for Craft Commerce
 
+## 3.0.0 - 2021-04-20
+
+### Changed
+- The plugin now requires Craft 3.6 and Commerce 3.3 or later.
+- The plugin now requires Guzzle 7.
+
+### Changed
+- Bank transfer payments with an `open`/`processing` status will now complete and order. ([#17](https://github.com/craftcms/commerce-mollie/issues/17))
+
 ## 2.1.2.1 - 2021-03-03
 
 ### Added
@@ -17,7 +26,7 @@
 
 ## 2.1.0.1 - 2019-07-24
 
-### Changed
+### Changed
 - Updated changelog with missing changes for 2.1.0
 
 ## 2.1.0 - 2019-07-24